generate function to return generated coupons
Hello there,

I had to use coupon generator and afterwards send out an email with the generated codes.

Doesn't it make more sense to return a structure containing the coupons instead of having to query back the
db for those?

This change works great for me but maybe it affects sylius in a way I cannot see at the moment. Please let me
know in case this can be achieved in any other way.

Cheers,
Argiris

adding @return hint on generate